feat: partner accepts Inquiry checkout phase 2 offer (#718)
* Add `defining_total?` method to offer

This indicates the offers that come after an offer without definite total with the same amount (if the amount is different it is a counter offer). It is added to model the scenatio where offers without inquiry checkout metadata are accepted after metadata is provided by the partner. We model this as a counter offer with the same ammount

* `SellerAcceptIndefiniteOffer` mutation

When after providing missing metadata, seller accepts buyer's offer on an artwork initially without shipping/location. This is captured on the backend as a counter offer with the same `amount_cents`.

* Complete integration tests for inquiry checkout phase 2

* Remove `isDefiningTotal`, add `feesAdded` and `offerAmountChanged`

* Rename `sellerAcceptIndefinite...` to `sellerAcceptProvisional...`

* Rename `feesChanged` to `definesTotal`

* Follow up on validation